allowance/requirements.txt
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytz==2023.3.post1